feat(web): unify connections + environments behind Group by toggle (#2460)
Collapse the parallel `/admin/connections` and `/admin/connections/groups`
admin pages into a single page with a `Group by: [Type | Environment]`
segmented control. Drops the "Environments" nav entry; the legacy URL
keeps working via a Next.js server-side `redirect()` to
`/admin/connections?groupBy=environment` so bookmarked deep links land on
the right grouping without a render flash.

The environment view embeds the existing connection-groups surface
verbatim — every capability (create / rename / delete / archive cascade
preview / merge wizard / auto-detected-singleton + archived toggles +
member roster) extracted into `ConnectionGroupsSection` and rendered
under the same shell as the type view. Same fetch surface
(`/api/v1/admin/connection-groups`), no new endpoint.

Grouping logic extracted into the pure `bucketizeConnections` module:
`(connections, dimension) => bucketized list` keyed by `dbType` or
`groupId`, with `null` / `undefined` `groupId` collapsing into a single
"no environment" bucket. Unit-testable in isolation; the page uses it
for both the existing provider-grouped layout and as the contract for
the new toggle. URL state via nuqs (`?groupBy=type|environment`).

Slice 4 of PRD #2458.
